chore: rearranged the package structure to support article feature
