+++ app/src/main/java/io/github/chhabra_dhiraj/spaceflightnews/feature/article/domain/util/Resource.kt
@@ -1,4 +1,4 @@
-package io.github.chhabra_dhiraj.spaceflightnews.domain.util
+package io.github.chhabra_dhiraj.spaceflightnews.feature.article.domain.util
 
 sealed class Resource<T>(val data: T? = null, val message: String? = null) {
     class Success<T>(data: T?): Resource<T>(data)
